Update values_crop.csv
I propose that value names should not contain spaces. Alternatively, we could add an additional column to contain the names without spaces. It is important to offer a name for tools that do not allow space (like ODK). If not, I'm afraid poeple will use different namings: _, - , snakecase, or anything else.